Read anndata file as the part of package
