+++ test/unit_test.py
@@ -14,6 +14,7 @@ if __name__ == "__main__":
         "test_backed_df.py",
         "test_backed_uns.py",
         "test_cap_anndata.py",
+        "test_reader.py",
     ]
     sys.exit(pytest.main(files))
